#WI19707:
Sonar code quality fixes

+++ src/main/java/com/mcmcg/ico/bluefin/model/PaymentProcessorMerchantResource.java
@@ -21,7 +21,7 @@ public class PaymentProcessorMerchantResource implements Serializable {
 
     public com.mcmcg.ico.bluefin.model.PaymentProcessorMerchant toPaymentProcessorMerchant() {
     	com.mcmcg.ico.bluefin.model.PaymentProcessorMerchant paymentProcessorMerchant = new com.mcmcg.ico.bluefin.model.PaymentProcessorMerchant();
-        paymentProcessorMerchant.setLegalEntityAppId(legalEntityAppId.longValue());
+        paymentProcessorMerchant.setLegalEntityAppId(legalEntityAppId);
         paymentProcessorMerchant.setMerchantId(merchantId);
         paymentProcessorMerchant.setTestOrProd(testOrProd);
